+++ src/components/SatelliteSearch.js
@@ -1,51 +1,21 @@
-import React, { useState, useEffect } from 'react';
+import React, { useRef } from 'react';
+import '../css/SatelliteSearch.css'
 
-const SatelliteSearch = () => {
-  const [satellites, setSatellites] = useState([]);
-  const [searchTerm, setSearchTerm] = useState('');
-  const [filteredSatellites, setFilteredSatellites] = useState([]);
+const SatelliteSearch = (props) => {
 
-  useEffect(() => {
-    const fetchSatellites = async () => {
-      try {
-        const response = await fetch('API_URL');// Need to replace this with Satellite API
-        const data = await response.json();
-        setSatellites(data);
-        setFilteredSatellites(data);
-      } catch (error) {
-        console.error('Error fetching satellite data:', error);
-      }
-    };
-
-    fetchSatellites();
-  }, []);
+  const searchRef = useRef();
 
   const handleSearch = (e) => {
-    const searchTerm = e.target.value.toLowerCase();
-    setSearchTerm(searchTerm);
-
-    const filtered = satellites.filter((satellite) =>
-      satellite.name.toLowerCase().includes(searchTerm)
-    );
-
-    setFilteredSatellites(filtered);
+    props.getSearchResult(searchRef.current.value.toLowerCase());
   };
 
   return (
-    <div>
-      <h1>Satellite Search</h1>
-      <input
-        type="text"
-        placeholder="Search for a satellite..."
-        value={searchTerm}
-        onChange={handleSearch}
-      />
-
-      <ul>
-        {filteredSatellites.map((satellite) => (
-          <li key={satellite.id}>{satellite.name}</li>
-        ))}
-      </ul>
+    <div className='SatelliteSearchBox'>
+      <h1 lass="display-6">Satellite Search</h1>
+      <div className="input-group mb-3">
+        <input type="text" placeholder="Search for a satellite..." className='form-control' ref={searchRef} />
+        <button type='button' className='btn btn-outline-secondary' onClick={handleSearch}>Search</button>
+      </div>
     </div>
   );
 };
