+++ src/configs/API_URL.js
@@ -1,0 +1,5 @@
+const APIs = {
+    fetch_all_satellites_api: 'API'
+};
+
+export default APIs;
